Bump version to 1.0.1 (#102)
I accidentally created a v1.0.0 tag for the 0.1.0 release. I have no
particular reason not to create a 1.0 version, so just do so now to
ensure consistency in version numbers and to ensure that the versions
are monotonically increasing everywhere.

Additionally, add a check to the publish step prior to actually
publishing.