Re-enable lint in CI (#100)
This change re-enables lint in CI.

I've updated some of the packages related to linting to bring them to
newer versions (not the latest eslint because foxglove eslint does not
support eslint v9 yet).

The extra tsconfig changes are to support linting the
`reflection.test.ts` file. Eslint wants the files it lints to also be
part of the typescript project so I've updates the `tsconfig.json` to
include all the `typescript` files in the `src` folder. However, I
_don't_ want the test file to be part of the build output so I have the
build-specific tsconfig files include only the entrypoint `index.ts`
file. During the build, `tsc` will traverse from the `index.ts` file and
include the files imported by it and not the test files.

---------

Co-authored-by: Jacob Bandes-Storch <jacob@bandes-stor.ch>