Fix union field table handling (#115)
This fixes a bug with union field deserialization. While
https://github.com/jkuszmaul/flatbuffers_reflection/pull/113 did fix the
buggy handling of the typeIndex for union discriminator field reading,
it did not address the same issue for `readTableLambda` and the use of
`readVectorOfTables` in union handling logic.

In both cases, the logic to prepare a new `Table` instance(s) to
correctly deserialize the union field was using an incorrect "field type
index". It was using the `fieldType.index()` value, which, for union
fields, is actually the index into the enums list for the types that
make up the union.

This change updates the union and union vector logic to use the correct
specific types (based on discriminator value) when creating a table
instance for further deserialization.

This change also surfaced that the union of struct handling code was
broken. To make it clear that union of struct is not supported, the
union of struct case is explicitly disallowed. A test is added showing
it is explicitly disallowed.