Change _null_ to _undefined_ when field value is not present (#116)
This change updates the return type of various deserialization functions
to return `undefined` (rather than null) when a field is not present.
This creates consistency with the javascript semantic that accessing a
field that does not exist will return undefined.

This change _does not_ change the behavior of _omitting_ these fields
from the resulting object. So if a field is not present in the
serialized message it is not present in the returned object.